# Use Attribute.GetCustomAttributes to reduce allocations / improve performance