@MattKotsenas 2024-08-15
If you're new to interpreting the Visual Studio allocation results, there's a good demo here walking through the process: https://youtu.be/TRFfTdzpk-M?si=gO4VrPjTi2EJTWLp